Adding last updates

+++ src/main/java/edu/pe/cibertec/SAIBM/service/impl/LibroServiceImpl.java
@@ -9,6 +9,8 @@ import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+
 @Service
 public class LibroServiceImpl extends GenericServiceImpl<LibroEntity, Integer> implements LibroService {
 
@@ -19,11 +21,20 @@ public class LibroServiceImpl extends GenericServiceImpl<LibroEntity, Integer> i
     private ReservaRepository reservaRepository;
 
     @Override
+    public List<LibroEntity> buscarPorNombre(String name) {
+        return libroRepository.findByNombreLibro(name);
+    }
+
+    @Override
     @Transactional
     public void eliminar(Integer id) {
-        reservaRepository.deleteByLibroId(id);
+        int countReservas = reservaRepository.countByLibroId(id);
+        if (countReservas > 0) {
+            throw new RuntimeException("No se puede eliminar el libro porque tiene reservas activas.");
+        }
         libroRepository.deleteById(id);
     }
+
 
     @PostConstruct
     private void init() {
